Move PlexServerConnection to plex module
